pyproject.toml
[project]
name = "customer_AI_agent"
version = "0.1.0"
description = "Categorize customer reviews feedback on Air France services using LLMs: detect sentiment (positive, negative, neutral), extract key topics and product mentions."
license = "MIT"
authors = [
    {name = "Mbogol Touye Achille", email = "touyejunior@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<=3.13"
dependencies = [
    "python-dotenv",
    "requests",
    "ffmpeg",
    "openai==1.76.2",
    "langchain==0.3.24",
    "langchain-core==0.3.56",
    "langchain-text-splitters==0.3.8",
    "langgraph==0.4.1",
    "langgraph-checkpoint==2.0.25",
    "langgraph-prebuilt==0.1.8",
    "langgraph-sdk==0.1.66",
    "langsmith==0.3.39",
    "langchain-community==0.3.23",
    "pydub==0.25.1"
]

[tool.poetry]
packages = [
    { include = "agent", from = "src" }
]

[tool.poetry.group.dev.dependencies]
black = "*"
flake8 = "*"
pre-commit = "*"
poetry = "*"
langgraph-cli="0.2.7"


[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"
